Fix mismatched type

+++ geometry/point.cc
@@ -25,7 +25,7 @@ lld cross(Point a, Point b) { return a.x*b.y - a.y*b.x; }
 lld ccw(Point a, Point b, Point c) { return cross(b-a, c-a); }
 double area(vector<Point> P) {
   lld sum = 0;
-  for (int i = 0; i < P.size()-1; i++) sum += cross(P[i], P[i+1]);
+  for (int i = 0; i < (int)P.size()-1; i++) sum += cross(P[i], P[i+1]);
   sum += cross(P.back(), P[0]);
   return abs(sum) / 2.;
 }
